Reducing Compiler Warnings
- Removed unused variables
- Introduced some casting
- Added `constexpr` to variables used as array sizing
- Cleaned up whitespace
- Changed uint_16 to regular int (this could cause overflow potentially)

+++ src/include/moveutil.h
@@ -12,6 +12,7 @@ The Multistrand Team (help@multistrand.org)
 #include <vector>
 #include <map>
 #include <sequtil.h>
+#include <cstdint>
 #include "basetype.h"
 
 using std::vector;
@@ -106,7 +107,7 @@ namespace std {
 template<> struct hash<ExportData> {
 	size_t operator()(const ExportData& k) const {
 
-		return hash<string>()(k.sequence) ^ hash<string>()(k.structure) ^ hash<int>()(k.id);
+		return hash<string>()(k.sequence) ^ hash<string>()(k.structure) ^ hash<double>()(k.id);
 
 	}
 };
@@ -114,7 +115,7 @@ template<> struct hash<ExportData> {
 template<> struct hash<ExportTransition> {
 	size_t operator()(const ExportTransition& k) const {
 
-		return hash<ExportData>()(k.state1) ^ hash<ExportData>()(k.state2) ^ hash<int>()(k.type);
+		return hash<ExportData>()(k.state1) ^ hash<ExportData>()(k.state2) ^ hash<double>()(k.type);
 
 	}
 };
